rename projectToPlane in helm* path3 plane* point* vec*
