+++ lib/helm2/index.js
@@ -126,8 +126,8 @@ exports.multiply = exports.compose
 exports.inverse = require('./invert')
 exports.invert = exports.inverse
 exports.limitDilation = require('./limitDilation')
-exports.projectTo = require('./projectTo')
-exports.projectToPlane = exports.projectTo
+exports.projectToPlane = require('./projectToPlane')
+exports.projectTo = exports.projectToPlane
 
 exports.scaleBy = require('./scaleBy')
