+++ lib/helm2/projectToPlane.js
@@ -1,7 +1,8 @@
 const transitTo = require('./transitTo')
 
 module.exports = (tr, plane, camera) => {
-  // @affineplane.helm2.projectTo(tr, plane[, camera])
+  // @affineplane.helm2.projectToPlane(tr, plane[, camera])
+  // @affineplane.helm2.projectTo
   //
   // Project transformation onto a plane.
   // If camera is given, project perspectively.
