+++ lib/helm3/index.js
@@ -113,8 +113,8 @@ exports.getTranslation = require('./getTranslation')
 exports.inverse = require('./invert')
 exports.invert = exports.inverse
 exports.limitDilation = require('./limitDilation')
-exports.projectTo = require('./projectTo')
-exports.projectToPlane = exports.projectTo
+exports.projectToPlane = require('./projectToPlane')
+exports.projectTo = exports.projectToPlane
 exports.rotateBy = require('./rotateBy')
 exports.scaleBy = require('./scaleBy')
 exports.setDilation = require('./setDilation')
