+++ lib/helm3/projectToPlane.js
@@ -1,7 +1,8 @@
 const transitTo = require('../helm2/transitTo')
 
 module.exports = (tr, plane) => {
-  // @affineplane.helm3.projectTo(tr, plane)
+  // @affineplane.helm3.projectToPlane(tr, plane)
+  // @affineplane.helm3.projectTo
   //
   // Project transformation onto a plane orthogonally.
   // The transformation loses z translation.
