+++ lib/path3/index.js
@@ -8,7 +8,7 @@
 
 exports.combine = require('./combine')
 exports.create = require('./create')
-exports.projectTo = require('./projectTo')
-exports.projectToPlane = exports.projectTo
+exports.projectToPlane = require('./projectToPlane')
+exports.projectTo = exports.projectToPlane
 exports.transitFrom = require('./transitFrom')
 exports.transitTo = require('./transitTo')
