+++ lib/path3/projectToPlane.js
@@ -1,8 +1,9 @@
 const transitTo = require('../path2/transitTo')
-const projectTo = require('../point3/projectTo')
+const projectTo = require('../point3/projectToPlane')
 
 module.exports = (path, plane, camera) => {
-  // @affineplane.path3.projectTo(path, plane[, camera])
+  // @affineplane.path3.projectToPlane(path, plane[, camera])
+  // @affineplane.path3.projectTo
   //
   // Project a 3D path onto a 2D plane orthogonally.
   // Project perspectively if a camera position is given.
