+++ lib/plane2/index.js
@@ -41,8 +41,8 @@ exports.getScale = require('./getScale')
 exports.inverse = require('./invert')
 exports.invert = exports.inverse
 exports.limitScale = require('./limitScale')
-exports.projectTo = require('./projectTo')
-exports.projectToPlane = exports.projectTo
+exports.projectToPlane = require('./projectToPlane')
+exports.projectTo = exports.projectToPlane
 exports.rotateBy = require('./rotateBy')
 exports.rotateTo = require('./rotateTo')
 exports.rotateToOrtho = require('./rotateToOrtho')
