+++ lib/plane2/projectToPlane.js
@@ -2,7 +2,8 @@ const scaleBy = require('./scaleBy')
 const transitTo = require('./transitTo')
 
 module.exports = (plane, target, camera) => {
-  // @affineplane.plane2.projectTo(plane, target[, camera])
+  // @affineplane.plane2.projectToPlane(plane, target[, camera])
+  // @affineplane.plane2.projectTo
   //
   // Project a 2D plane from reference to the target parallel 2D plane in 3D.
   // If camera is given, project perspectively, otherwise orthogonally.
