+++ lib/plane3/index.js
@@ -41,9 +41,9 @@ exports.getScale = require('./getScale')
 exports.inverse = require('./invert')
 exports.invert = exports.inverse
 exports.limitScale = require('./limitScale')
-exports.projectTo = require('./projectTo')
+exports.projectToPlane = require('./projectToPlane')
+exports.projectTo = exports.projectToPlane
 exports.projectToDepth = require('./projectToDepth')
-exports.projectToPlane = exports.projectTo
 exports.projectToScale = require('./projectToScale')
 exports.rotateBy = require('./rotateBy')
 exports.rotateTo = require('./rotateTo')
