+++ lib/plane3/projectToPlane.js
@@ -2,7 +2,8 @@ const scaleBy = require('../plane2/scaleBy')
 const transitTo = require('../plane2/transitTo')
 
 module.exports = (plane, target, camera) => {
-  // @affineplane.plane3.projectTo(plane, target[, camera])
+  // @affineplane.plane3.projectToPlane(plane, target[, camera])
+  // @affineplane.plane3.projectTo
   //
   // Project the plane based on the image plane position
   // and the optional camera position.
