+++ lib/point2/index.js
@@ -53,9 +53,9 @@ exports.offset = require('./offset')
 
 exports.polarOffset = require('./polarOffset')
 
-exports.projectTo = require('./projectTo')
+exports.projectToPlane = require('./projectToPlane')
+exports.projectTo = exports.projectToPlane
 exports.projectToLine = require('./projectToLine')
-exports.projectToPlane = exports.projectTo
 
 exports.rotateBy = require('./rotateBy')
