+++ lib/point2/projectToPlane.js
@@ -2,7 +2,8 @@ const transitTo = require('./transitTo')
 const homothety = require('./homothety')
 
 module.exports = (point, plane, camera) => {
-  // @affineplane.point2.projectTo(point, plane[, camera])
+  // @affineplane.point2.projectToPlane(point, plane[, camera])
+  // @affineplane.point2.projectTo
   //
   // Project a point onto another plane in 3d.
   // If camera is given, project perspectively.
