+++ lib/point3/index.js
@@ -22,8 +22,8 @@ exports.homothety = require('./homothety')
 exports.mean = exports.average
 exports.offset = require('./offset')
 exports.polarOffset = require('./polarOffset')
-exports.projectTo = require('./projectTo')
-exports.projectToPlane = exports.projectTo
+exports.projectToPlane = require('./projectToPlane')
+exports.projectTo = exports.projectToPlane
 exports.rotateBy = require('./rotateBy')
 exports.round = require('./round')
 exports.toArray = require('./toArray')
