+++ lib/point3/projectToPlane.js
@@ -2,7 +2,8 @@ const homothety = require('./homothety')
 const transitTo = require('../point2/transitTo')
 
 module.exports = (point, plane, camera) => {
-  // @affineplane.point3.projectTo(point, plane[, camera])
+  // @affineplane.point3.projectToPlane(point, plane[, camera])
+  // @affineplane.point3.projectTo
   //
   // Project a 3D point onto a plane in 3D space.
   //
