+++ lib/vec2/index.js
@@ -59,8 +59,8 @@ exports.opposite = exports.negation
 
 // exports.polarAverage
 
-exports.projectTo = require('./projectTo')
-exports.projectToPlane = exports.projectTo
+exports.projectToPlane = require('./projectToPlane')
+exports.projectTo = exports.projectToPlane
 exports.projectToVector = require('./projectToVector')
 
 exports.rotateBy = require('./rotateBy')
