+++ lib/vec2/projectToPlane.js
@@ -2,7 +2,8 @@ const scaleBy = require('./scaleBy')
 const transitTo = require('./transitTo')
 
 module.exports = (v, plane, camera) => {
-  // @affineplane.vec2.projectTo(v, plane[, camera])
+  // @affineplane.vec2.projectToPlane(v, plane[, camera])
+  // @affineplane.vec2.projectTo
   //
   // Project a vector onto another plane.
   // If camera is given, project perspectively.
