+++ lib/vec3/index.js
@@ -26,8 +26,8 @@ exports.magnitude = require('./magnitude')
 exports.negate = exports.invert
 exports.norm = exports.magnitude
 exports.normalize = require('./unit')
-exports.projectTo = require('./projectTo')
-exports.projectToPlane = exports.projectTo
+exports.projectToPlane = require('./projectToPlane')
+exports.projectTo = exports.projectToPlane
 exports.projectToVector = require('./projectToVector')
 exports.rotateAroundAxis = require('./rotateAroundAxis')
 exports.rotateBy = require('./rotateBy')
