+++ lib/vec3/projectToPlane.js
@@ -1,7 +1,8 @@
 const transitTo = require('../vec2/transitTo')
 
 module.exports = (vec, plane) => {
-  // @affineplane.vec3.projectTo(vec, plane)
+  // @affineplane.vec3.projectToPlane(vec, plane)
+  // @affineplane.vec3.projectTo
   //
   // Project a 3D vector onto a 2D plane orthogonally.
   // We cannot project 3D vectors perspectively because
