+++ test/plane3/index.test.js
@@ -11,7 +11,7 @@ const units = {
   getScale: require('./getScale.test'),
   invert: require('./invert.test'),
   limitScale: require('./limitScale.test'),
-  projectTo: require('./projectTo.test'),
+  projectToPlane: require('./projectToPlane.test'),
   projectToDepth: require('./projectToDepth.test'),
   projectToScale: require('./projectToScale.test'),
   rotateBy: require('./rotateBy.test'),
